routing: export from google3
* split test and benchmark of fourier solver
* use of StrongIndexMap
